pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MapCreator"
version = "0.1.2"
description = "A Python package for creating and visualizing maps."
readme = "README.md"
authors = [{ name = "Nicholas Orsini", email = "norsini1515@gmai.com" }]
requires-python = ">=3.11"
classifiers = ["Programming Language :: Python :: 3", "Operating System :: OS Independent"]
dependencies = [
  "numpy",
  "matplotlib",
  "geopandas",
  "shapely",
  "pillow",
  "opencv-python",
  "fiona",
  "typer>=0.12",      # <-- CLI
  "PyYAML>=6.0",      # <-- --config support
  "rasterio>=1.3",    # <-- raster init
  "affine>=2.4"       # <-- pixel->world transform
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mapcreator*"]   # <-- include subpackages like mapcreator.cli

[tool.setuptools.package-data]
mapcreator = ["cli/*.yml"]

[project.scripts]
mapc = "mapcreator.cli.__main__:app"
